agent/requirements.txt
google-generativeai>=0.3.0
google-genai>=0.3.0
ag-ui-adk>=0.1.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
